+++ wlc/main.py
@@ -25,6 +25,7 @@ import logging
 import sys
 from argparse import ArgumentParser
 from datetime import datetime
+from pathlib import Path
 from typing import Callable, Dict, List
 
 import argcomplete
@@ -631,11 +632,14 @@ class Unlock(ComponentCommand):
 
 
 @register_command
-class Download(TranslationCommand):
+class Download(ObjectCommand):
     """Downloads translation file."""
 
     name = "download"
-    description = "Downloads translation file"
+    description = """
+    Downloads translation file for a project, a component or a file
+    For a project or a component we will download a zip to --output
+    """
 
     @classmethod
     def add_parser(cls, subparser):
@@ -645,19 +649,79 @@ class Download(TranslationCommand):
             "-c", "--convert", help="Convert file format on server (defaults to none)"
         )
         parser.add_argument(
-            "-o", "--output", help="File where to store output (defaults to stdout)"
+            "-o",
+            "--output",
+            help="File|Directory where to store output (defaults to stdout)",
+        )
+        parser.add_argument(
+            "-g",
+            "--no-glossary",
+            action="store_true",
+            default=False,
+            help="Disable download glossary (defaults to False)",
         )
         return parser
 
+    def download_component(self, component):
+        """
+        Download a single component as file (if not a translation).
+        """
+
+        content = component.download(self.args.convert)
+        if self.args.output is None:
+            raise CommandError("Output is needed for download!")
+
+        directory = Path(self.args.output)
+        file_path = directory.joinpath(f"{component.project.slug}-{component.slug}.zip")
+
+        if not directory.exists():
+            directory.mkdir(exist_ok=True, parents=True)
+
+        with open(file_path, "wb") as file:
+            file.write(content)
+
+    def download_components(self, iterable):
+        for component in iterable:
+            # Ignore glossary via --no-glossary
+            if getattr(component, "is_glossary", False) and self.args.no_glossary:
+                continue
+            self.download_component(component)
+            item = f"{component.project.slug}/{component.slug}"
+            self.println(f"downloaded translations for component: {item}")
+
     def run(self):
         """Executor."""
-        obj = self.get_object()
-        content = obj.download(self.args.convert)
-        if self.args.output and self.args.output != "-":
-            with open(self.args.output, "wb") as handle:
-                handle.write(content)
-        else:
-            self.stdout.buffer.write(content)
+        obj = self.get_object(blank=True)
+
+        # Translation locale for a component
+        if isinstance(obj, wlc.Translation):
+            content = obj.download(self.args.convert)
+            if self.args.output and self.args.output != "-":
+                with open(self.args.output, "wb") as handle:
+                    handle.write(content)
+            else:
+                self.stdout.buffer.write(content)
+            return
+
+        # All translations for a component
+        if isinstance(obj, wlc.Component):
+            for component in self.wlc.list_components():
+                # Only download for the component we scoped
+                if obj.slug == component.slug:
+                    self.download_component(component)
+                    self.println(
+                        f"downloaded translations for component: {self.args.object[0]}"
+                    )
+
+            return
+
+        # All translations for a project
+        if isinstance(obj, wlc.Project):
+            self.download_components(obj.list())
+            self.println(f"downloaded translations for project: {self.args.object[0]}")
+            return
+
+        self.download_components(self.wlc.list_components())
 
 
 @register_command